+++ fortpy/scripts/runtests.py
@@ -1,28 +1,26 @@
 #!/usr/bin/env python
-from termcolor import cprint
 import argparse
-
-def print_result(testkey, percent, time, common):
+def print_result(testkey, percent, time, common, compiler):
     """Prints the specified result to the terminal with coloring based
     on how successful it was.
     """
     printkey = testkey.replace("|", " | ")
     if percent > .99:
-        color = "green"
+        pfun = lambda m: msg.okay(m, level=0)
     elif percent > .50:
-        color = "yellow"
+        pfun = lambda m: msg.warn(m, level=0, prefix=False)
     else:
-        color = "red"
+        pfun = lambda m: msg.err(m, level=0, prefix=False)
 
-    text = "RESULT: {0} \n\t{1:.2%} success ({3:.2%} common) in {2} ms\n"
-    cprint(text.format(printkey, percent, time, common), color)
+    text = "RESULT: {0} | {4} \n\t{1:.2%} success ({3:.2%} common) in {2} ms\n"
+    pfun(text.format(printkey, percent, time, common, compiler))
 
 def _get_compilers():
     """Returns a list of compilers from the command-line arguments."""
     complist = []
-    if args["compiler"] and args["compiler"] != "*":
-        complist.append(args["compiler"])
-    elif args["compiler"] == "*":
+    if args["compiler"] and args["compiler"][0] != "*":
+        complist.extend(args["compiler"])
+    elif args["compiler"] and args["compiler"][0] == "*":
         from fortpy.testing.compilers import compilers
         for c in compilers:
             complist.append(c)
@@ -32,13 +30,18 @@ def _get_compilers():
     
 def do_testing(args):
     """Runs the unit tests for all the modules in the code directory."""
+    from os import path
+    if not path.isdir(args["codedir"]):
+        msg.err("The source code directory '{}' does not exist.".format(args["codedir"]))
+        exit(2)
+    
     from fortpy.msg import set_quiet
     from fortpy.testing.tester import UnitTester
     set_quiet(args["quiet"])
     
     t = UnitTester(args["stagedir"], args["verbose"], args["templates"], args["fortpy"],
                    args["rerun"], debug=(not args["nodebug"]), profile=args["profile"],
-                   strict=args["strict"], quiet=args["quiet"])
+                   strict=args["strict"], quiet=args["quiet"], nprocs=args["nprocs"])
 
     #We only have to write the testing folder once; it gets copied for all
     #the remaining tests that need to be run for different compilers.
@@ -47,6 +50,7 @@ def do_testing(args):
     complist = _get_compilers()
     totalperc = 0
     totaltest = 0
+    msg.info("Running for compilers: {}".format(', '.join(complist)))
     for c in complist:
         result = t.runall(c)
         print("")
@@ -57,24 +61,24 @@ def do_testing(args):
                 timestr = "{0:.4f}".format(totaltime*1000)
             else:
                 timestr = "<untimed>"
-            print_result(idk, result[idk].percent, timestr, result[idk].common)
+            print_result(idk, result[idk].percent, timestr, result[idk].common, c)
             totalperc += result[idk].percent
             totaltest += 1
 
     #This section for exit codes helps the continuous integration server to know what's
     #going on with all the test results.
-
     if totaltest == 0 and totalperc == 0:
         _exit_code(0, "No Tests")
     else:
         score = totalperc/totaltest
-        
+
+    print("@CI: {0:.2%}".format(score))        
     if score == 1.:
         _exit_code(0, "Success")
     elif score < 1.:
-        _exit_code(2, "Failure")
+        _exit_code(4, "Failure")
     else:
-        _exit_code(3, "Didn't Run")
+        _exit_code(5, "Didn't Run")
 
 def _exit_code(i, prefix):
     """Informs the user of the exit code and then exits."""
@@ -129,7 +133,8 @@ parser.add_argument("-rerun",
                           "When a test is rerun, it is recompiled and tested, even if the code base "
                           "has not changed since the last test. Value '*' reruns the unit "
                           "tests of *all* modules in the code directory."))
-parser.add_argument("-compiler", help="Specify the compiler to use for the unit testing")
+parser.add_argument("-compiler", nargs="+",
+                    help="Specify the compiler(s) to use for the unit testing")
 parser.add_argument("-pypath", help="Specify a path to add to sys.path before running the tests.")
 parser.add_argument("-nodebug", 
                     help=("Compile the executables with DEBUG=false; the default behavior is to "
@@ -147,6 +152,11 @@ parser.add_argument("-strict", action="store_true",
 parser.add_argument("-compileaux", action="store_true",
                     help=("Also compile the fpy_auxiliary.f90 into .o, .mod and .so library. "
                           "Requires -stagedir to be specified."))
+parser.add_argument("-nocolor", action="store_true",
+                    help=("Don't output using termcolor. Useful when redirecting stdout so that "
+                          "it is easier to read."))
+parser.add_argument("-nprocs", type=int, default=1,
+                    help="Specify the number of processors to use for parallel execution.")
 
 if __name__ == "__main__":
     #Parse the args from the commandline that ran the script, call initialize
@@ -166,6 +176,8 @@ if __name__ == "__main__":
 
     from fortpy import msg
     msg.set_verbosity(args["verbose"])
+    if args["nocolor"]:
+        msg.nocolor = True
         
     testing = not (args["auxiliary"])
     if testing: